KL-145/chore: centralize stores
- store를 한군데로 모았습니다
